mios32_ain: deadband now with 16bit resolution

+++ mios32/LPC17xx/mios32_ain.c
@@ -107,7 +107,7 @@
 
 static u16 adc_conversion_values[NUM_CHANNELS_MAX];
 
-static u8 ain_deadband = MIOS32_AIN_DEADBAND;
+static u16 ain_deadband = MIOS32_AIN_DEADBAND;
 
 static u8  num_used_channels;
 
@@ -310,7 +310,7 @@ s32 MIOS32_AIN_DeadbandGet(void)
 //! be achieved to trigger the callback function passed to AINSER_Handler()
 //! \return < 0 on error
 /////////////////////////////////////////////////////////////////////////////
-s32 MIOS32_AIN_DeadbandSet(u8 deadband)
+s32 MIOS32_AIN_DeadbandSet(u16 deadband)
 {
 #if !MIOS32_AIN_CHANNEL_MASK
   return -1; // no analog input selected
